# Refactor jacobian and grad