--- a/src/FiniteDifferences.jl
+++ b/src/FiniteDifferences.jl
@@ -6,8 +6,6 @@ module FiniteDifferences
 
     include("methods.jl")
     include("numerics.jl")
+    include("to_vec.jl")
     include("grad.jl")
-
-
-    @deprecate jacobian(fdm, f, x::Vector, D::Int) jacobian(fdm, f, x; len=D)
 end