--- a/src/FiniteDifferences.jl
+++ b/src/FiniteDifferences.jl
@@ -2,8 +2,6 @@ module FiniteDifferences
 
     using Printf, LinearAlgebra
 
-    const AV = AbstractVector
-
     include("methods.jl")
     include("numerics.jl")
     include("to_vec.jl")