--- a/src/FiniteDifferences.jl
+++ b/src/FiniteDifferences.jl
@@ -2,6 +2,8 @@ module FiniteDifferences
 
     using Printf, LinearAlgebra
 
+    export to_vec, grad, jacobian, jvp, j′vp
+
     include("methods.jl")
     include("numerics.jl")
     include("to_vec.jl")