--- a/src/grad.jl
+++ b/src/grad.jl
@@ -17,7 +17,7 @@ function jacobian(fdm, f, x::Vector{<:Number}; len=nothing)
             xn = x[n]
             x[n] = xn + ε
             ret = first(to_vec(f(x)))
-            x[n] = xn  # Can't do `x[k] -= ϵ` as floating-point math is not associative
+            x[n] = xn  # Can't do `x[n] -= ϵ` as floating-point math is not associative
             return ret
         end
     end