--- a/src/to_vec.jl
+++ b/src/to_vec.jl
@@ -1,3 +1,5 @@
+export to_vec
+
 """
     to_vec(x)