--- a/src/to_vec.jl
+++ b/src/to_vec.jl
@@ -3,7 +3,7 @@ export to_vec
 """
     to_vec(x)
 
-Transform `x` into a `Vector`, and return a closure which inverts the transformation.
+Transform `x` into a `Vector`, and return the vector, and a closure which inverts the transformation.
 """
 function to_vec(x::Number)
     function Number_from_vec(x_vec)