This is a fairly substantial PR, motivated in no small part by @Roger-luo 's work in #54 .

### Overview

In particular it

1. completely refactors `jacobian` for performance reasons. Our previous implementation was really quite bad (my fault) -- it required ~~O(input_dims^2)~~ O(input_dims * output_dims) function evaluations rather than O(input_dims).
2. refactors `grad` to utilise `jpvp`. `grad` is now a single-line function that utilises the new `jacobian` implementation. This is a substantial drop in code complexity with no loss of functionality, as much of the code that we had before to handle dictionaries etc is completely gone. The result is that `grad` now depends on `to_vec` like everything else in this package.
3. the new implementation of `jacobian` has the pleasing advantage of doing away with the need to pass a `len` parameter around. This is a breaking change, hence the minor version bump.
4. `to_vec` has been moved out of `grad.jl` to emphasise the separation of concerns.

### Benchmarking

```julia
using BenchmarkTools, FiniteDifferences

A = randn(15, 10)
B = randn(10, 12)

@benchmark jacobian($(central_fdm(5, 1)), *, $A, $B)

# master
BenchmarkTools.Trial:
  memory estimate:  1.02 GiB
  allocs estimate:  4520890
  --------------
  minimum time:     637.381 ms (2.42% GC)
  median time:      654.539 ms (2.73% GC)
  mean time:        660.101 ms (2.80% GC)
  maximum time:     691.597 ms (3.33% GC)
  --------------
  samples:          8
  evals/sample:     1

# New code
BenchmarkTools.Trial:
  memory estimate:  15.97 MiB
  allocs estimate:  45663
  --------------
  minimum time:     5.207 ms (0.00% GC)
  median time:      5.380 ms (0.00% GC)
  mean time:        5.627 ms (4.77% GC)
  maximum time:     8.298 ms (13.32% GC)
  --------------
  samples:          888
  evals/sample:     1
```