@willtebbutt 2020-03-31
@oxinabox thanks for the above comments. They're now resolved other than the deprecation, which I would appreciate feedback on.