@oxinabox 2020-03-31
> They're now resolved other than the deprecation, which I would appreciate feedback on

Have given feedback.

Ideally would wait for @Roger-luo  to give feedback as to if this worksout, before we merge this